First iteration of a very wrong shadow map. =)

+++ 18Lights/vklearnin/src/rendering/renderpass/RenderPass.cpp
@@ -8,49 +8,80 @@
 namespace vkl {
 
 // =============================================================================
-void RenderPass::create() {
-    _find_depth_stencil_format();
-    _init_color_buffer();
-    _init_depth_buffer();
-    _default_attachments();
-    _default_subpasses();
+RenderPass & RenderPass::init_color_buffer(uint32_t const width,
+                                           uint32_t const height)
+{
+    if(_color_buffer.handle) {
+        ImageTools::destroy(_color_buffer);
+    }
 
-    const vk::RenderPassCreateInfo renderpass_info {
-        .attachmentCount = static_cast<uint32_t>(_attach_descs.size()),
-        .pAttachments    = _attach_descs.data(),
-        .subpassCount    = static_cast<uint32_t>(_subpasses.size()),
-        .pSubpasses      = _subpasses.data(),
-        .dependencyCount = static_cast<uint32_t>(_subpass_deps.size()),
-        .pDependencies   = _subpass_deps.data()
+    _color_buffer.format = Swapchain::image_format();
+    _color_buffer.extent = vk::Extent3D {
+        .width  = width,
+        .height = height,
+        .depth  = 1u
     };
 
-    _render_pass = LogicalDevice::native().createRenderPass(renderpass_info);
-
-    CONSOLE_TRACE(
-        "Created Render Pass {:#x}",
-        reinterpret_cast<uint64_t>(VkRenderPass(_render_pass))
+    ImageTools::create(
+        _color_buffer,
+        vk::ImageType::e2D,
+        RenderConfig::max_msaa_flag(),
+        (
+            vk::ImageUsageFlagBits::eColorAttachment |
+            vk::ImageUsageFlagBits::eTransientAttachment
+        ),
+        vk::MemoryPropertyFlagBits::eDeviceLocal
     );
+
+    ImageTools::create_view(
+        _color_buffer,
+        vk::ImageViewType::e2D,
+        vk::ImageAspectFlagBits::eColor
+    );
+
+    return *this;
 }
 
 // =============================================================================
-void RenderPass::destroy() {
-    CONSOLE_TRACE(
-        "Destroying Render Pass {:#x}",
-        reinterpret_cast<uint64_t>(VkRenderPass(_render_pass))
+RenderPass & RenderPass::init_depth_buffer(uint32_t const width,
+                                           uint32_t const height)
+{
+    if(_depth_buffer.handle) {
+        ImageTools::destroy(_depth_buffer);
+    }
+
+    _find_depth_stencil_format();
+
+    _depth_buffer.format = _depth_format;
+    _depth_buffer.extent = vk::Extent3D {
+        .width  = width,
+        .height = height,
+        .depth  = 1u
+    };
+
+    ImageTools::create(
+        _depth_buffer,
+        vk::ImageType::e2D,
+        RenderConfig::max_msaa_flag(),
+        vk::ImageUsageFlagBits::eDepthStencilAttachment,
+        vk::MemoryPropertyFlagBits::eDeviceLocal
     );
 
-    ImageTools::destroy(_depth_buffer);
-    ImageTools::destroy(_color_buffer);
-    LogicalDevice::native().destroyRenderPass(_render_pass);
-    _render_pass = nullptr;
+    ImageTools::create_view(
+        _depth_buffer,
+        vk::ImageViewType::e2D,
+        vk::ImageAspectFlagBits::eDepth
+    );
+
+    return *this;
 }
 
 // =============================================================================
-void RenderPass::_default_attachments() {
+RenderPass & RenderPass::default_color_attachments() {
     _attach_descs = {{
         // color buffer (msaa) attachment description
         .format         = Swapchain::image_format(),
-        .samples        = _samples,
+        .samples        = RenderConfig::max_msaa_flag(),
         .loadOp         = vk::AttachmentLoadOp::eClear,
         .storeOp        = vk::AttachmentStoreOp::eDontCare,
         .stencilLoadOp  = vk::AttachmentLoadOp::eDontCare,
@@ -60,13 +91,13 @@ void RenderPass::_default_attachments() {
     },
     {   // depth buffer attachment description
         .format         = _depth_buffer.format,
-        .samples        = _samples,
+        .samples        = RenderConfig::max_msaa_flag(),
         .loadOp         = vk::AttachmentLoadOp::eClear,
         .storeOp        = vk::AttachmentStoreOp::eDontCare,
         .stencilLoadOp  = vk::AttachmentLoadOp::eDontCare,
         .stencilStoreOp = vk::AttachmentStoreOp::eDontCare,
         .initialLayout  = vk::ImageLayout::eUndefined,
-        .finalLayout = vk::ImageLayout::eDepthStencilAttachmentOptimal,
+        .finalLayout    = vk::ImageLayout::eDepthStencilAttachmentOptimal,
     },
     {   // final presentation attachment
         .format         = Swapchain::image_format(),
@@ -93,35 +124,12 @@ void RenderPass::_default_attachments() {
         .attachment = 2u,
         .layout = vk::ImageLayout::eColorAttachmentOptimal,
     }};
+
+    return *this;
 }
 
 // =============================================================================
-void RenderPass::_default_subpasses() {
-    _subpasses = {{
-        // This subpass is a graphical one
-        .pipelineBindPoint = vk::PipelineBindPoint::eGraphics,
-
-        // ...Which has no input of any kind
-        .inputAttachmentCount = 0u,
-        .pInputAttachments    = nullptr,
-
-        // But does have a single color attachment
-        .colorAttachmentCount =
-            static_cast<uint32_t>(_color_attachments.size()),
-        .pColorAttachments   = _color_attachments.data(),
-
-        // With whatever MSAA samples we've got
-        .pResolveAttachments = _resolve_attachments.data(),
-
-        // With a depth stencil
-        .pDepthStencilAttachment = &_depth_attachment,
-
-        // As we've only got a single subpass, there's nothing to preserve
-        // between subpasses
-        .preserveAttachmentCount = 0u,
-        .pPreserveAttachments    = nullptr,
-    }};
-
+RenderPass & RenderPass::default_color_subpass() {
     _subpass_deps = {{
         .srcSubpass      = VK_SUBPASS_EXTERNAL,
         .dstSubpass      = 0u,
@@ -142,66 +150,132 @@ void RenderPass::_default_subpasses() {
         .dstAccessMask   = vk::AccessFlagBits::eDepthStencilAttachmentWrite,
         .dependencyFlags = vk::DependencyFlagBits::eByRegion
     }};
+
+    _subpasses = {{
+        // This subpass is a graphical one
+        .pipelineBindPoint = vk::PipelineBindPoint::eGraphics,
+
+        // ...Which has no input of any kind
+        .inputAttachmentCount = 0u,
+        .pInputAttachments    = nullptr,
+
+        // But does have a single color attachment
+        .colorAttachmentCount =
+            static_cast<uint32_t>(_color_attachments.size()),
+        .pColorAttachments = _color_attachments.data(),
+
+        // With whatever MSAA samples we've got
+        .pResolveAttachments = _resolve_attachments.data(),
+
+        // With a depth stencil
+        .pDepthStencilAttachment = &_depth_attachment,
+
+        // As we've only got a single subpass, there's nothing to preserve
+        // between subpasses
+        .preserveAttachmentCount = 0u,
+        .pPreserveAttachments    = nullptr,
+    }};
+
+    return *this;
 }
 
 // =============================================================================
-void RenderPass::_init_color_buffer() {
+RenderPass & RenderPass::init_shadow_map(uint32_t const resolution) {
+    _shadow_map.destroy();
+
+    _find_depth_stencil_format();
+    _shadow_map.create_shadow_map(resolution, resolution, _depth_format);
+
+    return *this;
+}
+
+// =============================================================================
+RenderPass & RenderPass::default_shadow_map_attachments() {
+    _attach_descs = {{
+        .format         = _depth_format,
+        .samples        = vk::SampleCountFlagBits::e1,
+        .loadOp         = vk::AttachmentLoadOp::eClear,
+        .storeOp        = vk::AttachmentStoreOp::eStore,
+        .stencilLoadOp  = vk::AttachmentLoadOp::eDontCare,
+        .stencilStoreOp = vk::AttachmentStoreOp::eDontCare,
+        .initialLayout  = vk::ImageLayout::eUndefined,
+        .finalLayout    = vk::ImageLayout::eDepthStencilReadOnlyOptimal,
+    }};
+
+    _shadow_map_attachment = {
+        .attachment = 0u,
+        .layout     = vk::ImageLayout::eDepthStencilAttachmentOptimal,
+    };
+
+    return *this;
+}
+
+// =============================================================================
+RenderPass & RenderPass::default_shadow_map_subpass() {
+    _subpass_deps = {{
+        .srcSubpass      = VK_SUBPASS_EXTERNAL,
+        .dstSubpass      = 0u,
+        .srcStageMask    = vk::PipelineStageFlagBits::eFragmentShader,
+        .dstStageMask    = vk::PipelineStageFlagBits::eEarlyFragmentTests,
+        .srcAccessMask   = vk::AccessFlagBits::eShaderRead,
+        .dstAccessMask   = vk::AccessFlagBits::eDepthStencilAttachmentWrite,
+        .dependencyFlags = vk::DependencyFlagBits::eByRegion
+    },
+    {
+        .srcSubpass      = 0u,
+        .dstSubpass      = VK_SUBPASS_EXTERNAL,
+        .srcStageMask    = vk::PipelineStageFlagBits::eLateFragmentTests,
+        .dstStageMask    = vk::PipelineStageFlagBits::eFragmentShader,
+        .srcAccessMask   = vk::AccessFlagBits::eDepthStencilAttachmentWrite,
+        .dstAccessMask   = vk::AccessFlagBits::eShaderRead,
+        .dependencyFlags = vk::DependencyFlagBits::eByRegion
+    }};
+
+    _subpasses = {{
+        .pipelineBindPoint       = vk::PipelineBindPoint::eGraphics,
+        .pDepthStencilAttachment = &_shadow_map_attachment,
+    }};
+
+    return *this;
+}
+
+// =============================================================================
+void RenderPass::create() {
+    const vk::RenderPassCreateInfo renderpass_info {
+        .attachmentCount = static_cast<uint32_t>(_attach_descs.size()),
+        .pAttachments    = _attach_descs.data(),
+        .subpassCount    = static_cast<uint32_t>(_subpasses.size()),
+        .pSubpasses      = _subpasses.data(),
+        .dependencyCount = static_cast<uint32_t>(_subpass_deps.size()),
+        .pDependencies   = _subpass_deps.data()
+    };
+
+    _render_pass = LogicalDevice::native().createRenderPass(renderpass_info);
+
+    CONSOLE_TRACE(
+        "Created Render Pass {:#x}",
+        reinterpret_cast<uint64_t>(VkRenderPass(_render_pass))
+    );
+}
+
+// =============================================================================
+void RenderPass::destroy() {
+    CONSOLE_TRACE(
+        "Destroying Render Pass {:#x}",
+        reinterpret_cast<uint64_t>(VkRenderPass(_render_pass))
+    );
+
+    if(_depth_buffer.handle) {
+        ImageTools::destroy(_depth_buffer);
+    }
     if(_color_buffer.handle) {
         ImageTools::destroy(_color_buffer);
     }
 
-    _samples = max_msaa_flag();
+    _shadow_map.destroy();
 
-    _color_buffer.format = Swapchain::image_format();
-    _color_buffer.extent = vk::Extent3D {
-        .width  = Swapchain::extent().width,
-        .height = Swapchain::extent().height,
-        .depth  = 1u
-    };
-
-    ImageTools::create(
-        _color_buffer,
-        vk::ImageType::e2D,
-        _samples,
-        (
-            vk::ImageUsageFlagBits::eColorAttachment |
-            vk::ImageUsageFlagBits::eTransientAttachment // ??????????????????????????????????????????????????????????????????
-        ),
-        vk::MemoryPropertyFlagBits::eDeviceLocal
-    );
-
-    ImageTools::create_view(
-        _color_buffer,
-        vk::ImageViewType::e2D,
-        vk::ImageAspectFlagBits::eColor
-    );
-}
-
-// =============================================================================
-void RenderPass::_init_depth_buffer() {
-    if(_depth_buffer.handle) {
-        ImageTools::destroy(_depth_buffer);
-    }
-
-    _depth_buffer.extent = vk::Extent3D {
-        .width  = Swapchain::extent().width,
-        .height = Swapchain::extent().height,
-        .depth  = 1u
-    };
-
-    ImageTools::create(
-        _depth_buffer,
-        vk::ImageType::e2D,
-        _samples,
-        vk::ImageUsageFlagBits::eDepthStencilAttachment,
-        vk::MemoryPropertyFlagBits::eDeviceLocal
-    );
-
-    ImageTools::create_view(
-        _depth_buffer,
-        vk::ImageViewType::e2D,
-        vk::ImageAspectFlagBits::eDepth
-    );
+    LogicalDevice::native().destroyRenderPass(_render_pass);
+    _render_pass = nullptr;
 }
 
 // =============================================================================
@@ -221,7 +295,7 @@ void RenderPass::_find_depth_stencil_format() {
                 "Using depth stencil format {}",
                 to_string(option)
             );
-            _depth_buffer.format = option;
+            _depth_format = option;
             return;
         }
     }
@@ -235,12 +309,13 @@ RenderPass::RenderPass() :
     _color_attachments   { },
     _depth_attachment    { },
     _resolve_attachments { },
-    _subpasses           { },
     _subpass_deps        { },
+    _subpasses           { },
+    _depth_format        { },
     _render_pass         { },
-    _samples             { },
     _color_buffer        { },
-    _depth_buffer        { }
+    _depth_buffer        { },
+    _shadow_map          { }
 { }
 
 } // namespace vkl
